Refactored mapper
