+++ app/api/core/infrastructure/mongodb/entity/MongoEntityMapper.ts
@@ -9,54 +9,7 @@ import { PropertyAssignment } from '#api/core/domain/template/PropertyValue.js';
 import { TemplateDBO } from '../template/DBOs/TemplateDBO.js';
 import { LoggerFactory } from '../../factories/LoggerFactory.js';
 import { MongoTemplateMapper } from '../template/MongoTemplateMapper.js';
-
-const isRecord = (value: unknown): value is Record<string, unknown> =>
-  !!value && typeof value === 'object' && !Array.isArray(value);
-
-const getIconRecord = (entry: unknown) => {
-  if (!isRecord(entry)) return null;
-  const { icon } = entry;
-  if (!isRecord(icon)) return null;
-  return icon;
-};
-
-const relationshipEntryToDBO = (entry: unknown) => {
-  if (!isRecord(entry)) return entry;
-  const iconRecord = getIconRecord(entry);
-  if (!iconRecord) return entry;
-
-  const iconId = iconRecord.id;
-  if (iconId === undefined) return entry;
-
-  const { id, ...restIcon } = iconRecord;
-
-  return {
-    ...entry,
-    icon: {
-      _id: id,
-      ...restIcon,
-    },
-  };
-};
-
-const relationshipEntryToDomain = (entry: unknown) => {
-  if (!isRecord(entry)) return entry;
-  const iconRecord = getIconRecord(entry);
-  if (!iconRecord) return entry;
-
-  const iconId = iconRecord._id ?? iconRecord.id;
-  if (iconId === undefined) return entry;
-
-  const { _id, id, ...restIcon } = iconRecord;
-
-  return {
-    ...entry,
-    icon: {
-      id: iconId,
-      ...restIcon,
-    },
-  };
-};
+import { MongoRelationshipMetadataMapper } from './MongoRelationshipMetadataMapper.js';
 
 class MongoEntityLanguageMapper {
   static toDomain(dbo: EntityDBO, template: Template): EntityTranslationProps {
@@ -80,7 +33,7 @@ class MongoEntityLanguageMapper {
 
       const mappedValue =
         property.getData().type === 'relationship' && Array.isArray(value)
-          ? value.map(entry => relationshipEntryToDomain(entry))
+          ? MongoRelationshipMetadataMapper.toDomain(value)
           : value;
 
       return {
@@ -143,7 +96,7 @@ class MongoEntityMapper {
         if (propertyValue.type === 'relationship') {
           return {
             ...acc,
-            [key]: propertyValue.value.map(entry => relationshipEntryToDBO(entry)),
+            [key]: MongoRelationshipMetadataMapper.toDBO(propertyValue.value),
           };
         }
